requirements.txt
# Multichannel ASR - Requirements
# Python 3.10+ required

# Deep Learning Framework
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Audio Processing
pydub>=0.25.1
webrtcvad>=2.0.10
soundfile>=0.12.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0

# Text Processing
opencc-python-reimplemented>=0.1.7

# Utilities
tqdm>=4.65.0

# Note: For CUDA 12.2+ users, install PyTorch with:
# pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu126
#
# For CPU-only installation:
# pip3 install torch torchvision
#
# System dependency (required for audio format conversion):
# - Ubuntu/Debian: sudo apt-get install ffmpeg
# - macOS: brew install ffmpeg
# - Windows: Download from https://ffmpeg.org/download.html
